requirement.txt
Django==1.7.3
Scrapy==0.24.4
Twisted==14.0.2
cffi==0.8.6
characteristic==14.3.0
cryptography==0.7.2
cssselect==0.9.1
django-haystack==2.3.1
elasticsearch==1.4.0
enum34==1.0.4
lxml==3.4.1
psycopg2==2.6
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pytz==2014.10
queuelib==1.2.2
service-identity==14.0.0
six==1.9.0
urllib3==1.10.2
w3lib==1.11.0
wsgiref==0.1.2
zope.interface==4.1.2
